extika v0.0.1 ExTika

Wrapper for Apache Tika, a toolkit that detects and extracts metadata and text from over a thousand different file types.

Functions

get_content_type(file)

Specs

get_content_type(String.t) :: String.t

Returns the content type of the document.

Examples

iex> ExTika.get_content_type("test/test-files/test.doc")
{:ok, "application/msword"}

get_content_type!(file)

Specs

get_content_type!(String.t) :: String.t

Returns the content type of the document. Fails on error.

Examples

iex> ExTika.get_content_type!("test/test-files/test.doc")
"application/msword"

get_language(file)

Specs

get_language(String.t) :: String.t

Returns the language of the document.

Examples

iex> ExTika.get_language("test/test-files/test.doc")
{:ok, "en"}

get_language!(file)

Specs

get_language!(String.t) :: String.t

Returns the language of the document. Fails on error.

Examples

iex> ExTika.get_language!("test/test-files/test.doc")
"en"

get_metadata(file)

Specs

get_metadata(String.t) ::
  {:ok, map} |
  {:error, String.t}

Fetches metadata about the document.

Examples

iex> {:ok, meta} = ExTika.get_metadata("test/test-files/test.docx")
iex> meta["meta:creation-date"]
"2016-08-14T14:51:38Z"

get_metadata!(file)

Specs

get_metadata!(String.t) :: map

Fetches metadata about the document. Fails on error.

Examples

iex> meta = ExTika.get_metadata!("test/test-files/test.docx")
iex> meta["meta:creation-date"]
"2016-08-14T14:51:38Z"

get_text(file)

Specs

get_text(String.t) ::
  {:ok, String.t} |
  {:error, String.t}

Extracts the text from the given file.

Examples

iex> ExTika.get_text("test/test-files/test.doc")
{:ok, "This is a DOC file.\n\n"}

get_text!(file)

Specs

get_text!(String.t) :: String.t

Extracts the text from the given file. Fails on error.

Examples

iex> ExTika.get_text!("test/test-files/test.doc")
"This is a DOC file.\n\n"